Renames StrongType.hpp to IndexingNamedType.hpp
