+++ PlayRho/Dynamics/Contacts/ContactID.hpp
@@ -21,7 +21,7 @@
 #ifndef PLAYRHO_DYNAMICS_CONTACTS_CONTACTID_HPP
 #define PLAYRHO_DYNAMICS_CONTACTS_CONTACTID_HPP
 
-#include <PlayRho/Common/StrongType.hpp>
+#include <PlayRho/Common/IndexingNamedType.hpp>
 #include <PlayRho/Common/Settings.hpp>
 
 namespace playrho {
